Add files via upload
A game of higher or lower. The program takes a list of dictionaries with information of celebrities including their Instagram follower count. The user is randomly given two celebrities from the list and has to guess which has the higher Instagram follower count. The program continues in a loop until the user gets it wrong.