Add files via upload
A coffee machine with a selection of drinks. Using functions and by returning values from functions the coffee machine checks to see if there are enough resources to produce the drink, asks the user for the coin input and makes the drink for the user if the user has paid enough money.