Add files via upload
The quiz project takes questions from one file and stores them as objects. In another class, I refer to these questions and output them to the user. The user then answers the question and is told whether they got the question correct or incorrect. At the end of the quiz the user is told their final score. This section was an exercise in getting used to using and writing classes for OOP.