Add files via upload
Snake part 1 is the first of two parts. This part was all about getting the snake on the screen and being able to move the snake up, left, down and right using the arrow keys on the keyboard. This was all done by using Turtle Graphics and utilising the documentation provided.