Add files via upload
The continuation of snake game part 1. In this section I brought other classes in to the code and used them for their respective functions. I was introduced to inheritance and how to utilise it when using OOP. The game now ends with the snake collides with the wall and when the snake collides with its own body. There is a score counter to keep track of the users score and the program outputs a game over message to the screen when the player dies.