Add files via upload
Animal crossing is a basic take on getting a turtle to cross the road and reach the other side. When the turtle gets to the other side, the speed of the cars increase until the turtle is finally hit by a car. This is done by utilising the turtle graphics module and OOP.