Add files via upload
Mail merge is a program which takes an invitation letter with a '[name]' place holder for the recipient and creates an individual letter for each person in the invited list which references the recipient by name. This exercise was about reading from and writing to files and getting used to the directory system when trying to find the required file.