Add files via upload
This project was about creating a miles to KM converter with a GUI. A miles to KM is simple enough to create but the main focus was to get this working with its own GUI. This was done by utilising the Tkinter module. With this, I was able to create labels, buttons and input with functions being called on button clicks. I was also able to lay it out nearly using grid.