Add files via upload
The Pomodoro Technique is a time management method developed by Francesco Cirillo in the late 1980s. This app sets timers for study and breaks based off the Pomodoro method. This is done by using Tkinter as a GUI and applying different countdowns depending on which rep the user is on.