Add files via upload
By utilising Tkinter, the password generator and storer is a GUI which prompts the user to enter a website name, and email address and a password. When submitted, this information will be written to a file and stored locally. When the user submits information they will be prompted to check what they have entered before confirming. If an input is not filled in, the user will be notified and prompted to complete the field. There is also a button that generates a random combination of letters, numbers and symbols for the user should they require it.