Add files via upload
This section was about reading from and writing to JSON files and exception handling. Expanding on the previous password generator project, the program now allows users to search a website name and if it exists in the file, the user's email and password will be shown to them in a popup box. The program now throws exceptions when a file is not found (file is then created)  and the program also lets the user know when the website information they are searching for doesn't exist. Overall the user experience is better because it doesn't keep them in the dark when errors and unsuccessful searches occur.