Add files via upload
The flash card app is a project that takes popular words in French and gives the user 3 seconds to guess the English equivalent (in their heads). The user is then shown the correct answer. From here, the user selects the X if they got it wrong and the checkmark if they got it right. If the user gets the word correct, that word will be removed from the list and a 'words_to_learn' file will be updated with all the words from the original file minus the words that have been correctly guessed.