Add files via upload
The birthday wisher is a program which reads from a file of birthdays and checks for if one of the birthdays on the list is today. This is done by utilising pandas to store the birthday list into a dictionary and import smtplib to email a happy birthday message if it is someone's birthday. To get the current date and month I imported datetime into the program and compared it to the dates in the birthdays file.