Add files via upload
The ISS project is a program that uses APIs to assess the current time, your personal longitude and latitude (user has to enter this), sunset time and the location of the International Space Station. By using the smtplib module, the program then emails the users (from their own email address) to notify them that the international space station is currently above them. However, the email is only sent after sunset.