Add files via upload
Revisiting a previous quiz project which involved OOP, this project was another quiz but this time with a GUI. By utilising Tkinter I was able to create a GUI for the user which displayed their score, the question and true or false buttons (as the questions were true or false questions). The questions are randomly chosen by using an API key from the open trivia database. If the user is correct, the screen will momentarily turn green, if the user is wrong, red. The user is notified when the quiz is over and presented with their score.