Add files via upload
The rain notification project sends a notification via email to myself every morning at 7am to let me know whether or not it will rain that day (in the next 12 hours). This is done by utilising an API key from openweathermap.org. From here I was able to use one of their API keys to dig through the dictionary or data and send a message to my own email (using smtplib) depending on whether or not it was going to rain that day, according to the weather.