Add files via upload
The stock notifications project uses API keys from both alphavantage and newsapi to get the values of the stocks of a certain company at the close of the trading day for yesterday and the day before. If the percentage change is over a certain threshold, then an email will be sent to myself to notify me of this change. Along with this change, in the email I will also receive the latest news story regarding the company in question.