Add files via upload
The habit tracker project is a program then sends post, put and delete requests to a service to take control of a habit tracker provided by Pixela. I create a username and token ID which enabled me to create my habit tracker graph and then use put and delete requests to update that graph to keep track of whichever particular habit I chose to track. The project was mainly about further learning how to read documentation and implement that to achieve certain results.