Add files via upload
The exercise log is a project that allows the user to log their exercise by using plain English. The user should be able to type in what exercise they did and that input will be compiled into several pieces of data and inserted into a google datasheet. For example, a user may enter "ran for 30 minutes" and from here, the date, time, exercise, duration and calories burned will be entered into their respective columns on a google sheet. This is done by utilising the nutritionix and sheety API's. On top of this, I used environmental variables to ensure the sensitive information would not be shown when project is published online.