Add files via upload
The cheap flights checker is a program that uses APIs to access information regarding flights for the foreseeable future. The program compares a google sheet of destinations with a max price. Using this max price I was able to return all flights to these destinations in the next 6 months that were lower than that max price. The idea is to get an email alert when a really good deal is found. The project isn't finished in its current state. It's been a fantastic project to practice on as I am getting a lot more experience with retrieving and processing data by using API's.