Add files via upload
Back to Python! This project was a lesson in web scraping which is essentially going through HTML code and picking out bits of information based on their HTML tags. In this project, I visited the empire website for the top 100 websites. I used web scraping to go through the HTML code and pick out the title of the top 100 films. I then reversed this list to start it from film number 1 to film number 100. I then wrote to a local file which didn't exist yet, the film and its ranking.