Add files via upload
The price tracker takes an item the user wants to track the price of from Amazon. Using web scraping, the program then grabs the price of the item and stores it as a float variable. The idea is to get this program to run daily automatically on software such as python anywhere and if the price is below a certain value set by the user, the user will receive an email to notify them that the product is on offer for a good price.