Add files via upload
The web bot is a project that utilises selenium (a python module) to allow the user to open up a website and navigate the website as a human would do with a mouse and keyboard. To demonstrate this, the program opens up the cookie clicker game (http://orteil.dashnet.org/experiments/cookie/) and repeatedly clicks the cookie to score more points. Every 10 seconds, the program will check what upgrade we can afford to buy and buys the most expensive upgrade available. This is done with similar methods to web scraping, checking the names of classes as they change throughout the game.