Add files via upload
This project was a program with an aim to get introduced to the very basic of flask and putting it into action. After setting up a virtual server using flask, I created a number guessing game. The user would guess a number via the end of the virtual URL which took them to a page. If that number was too low or too high, the user would be informed until they got the correct guess via the URL.