Add files via upload
This project is a program that utilises flask to make a local network to host html code. I have done this previously however this section was about bring HTML templates into my python code with flask but then passing arguments that could then be used on the HTML files with the correct syntax.